subt: Move Trace to separate file.
